chore: 개발 환경 설정 최적화 및 Swagger 모듈화 (#31)
* refactor: Swagger 설정 로직을 별도 파일로 분리하여 코드 가독성 향상

* ci: CI/CD 워크플로우 최적화

- Node 22 버전 적용, API 시크릿 추가 및 오토 라벨러 트리거(edited) 최적화

* docs: 프로젝트 문서 및 템플릿 업데이트

- 최신 디렉토리 구조 반영 및 효율적인 AI 리뷰 활용을 위한 주의사항 추가

* chore: 커밋 규칙 및 라벨 환경 설정

- init 타입 제거, subject-case 무시 설정 및 GitHub 라벨 싱크 최적화
