+++ .github/pull_request_template.md
@@ -1,7 +1,7 @@
 ## PR 체크리스트
 
 > [!TIP]
-> 자세한 협업 규칙은 [README 가이드라인](https://github.com/team-walker/backend#-협업-가이드)을 참고해 주세요!
+> 자세한 협업 규칙은 [협업 가이드라인](https://www.notion.so/hayeonbaek/2fdf2cf9d94180f488aef3da85e6e993?source=copy_link)을 참고해 주세요!
 
 **✅ 리뷰 요청 전, 아래 규칙 한 번 더 확인해 주세요!**
 
@@ -71,10 +71,6 @@
 
 <!-- API 테스트 결과, DB 스키마, 서버 로그 등 참고할 만한 스크린샷이 있다면 첨부해 주세요 -->
 
-### Before
-
-### After
-
 ## 📚 참고 자료 (선택사항)
 
 <!-- 관련 문서, 링크, 레퍼런스 등이 있다면 적어주세요 -->
@@ -82,9 +78,15 @@
 -
 
 ## 🤖 AI 코드 리뷰 활용 팁
+> [!WARNING]
+>
+> **⚠️ 효율적인 리뷰 워크플로우**
+>
+> `/gemini review`를 너무 자주 요청하면 AI가 반복적으로 수정 제안을 계속할 수 있어 오히려 비효율적일 수 있습니다. 불필요한 코멘트로 PR이 길어지지 않도록, 대규모 변경이 있을 때나 최종 확인 단계에서 활용하시는 것을 권장합니다.
 
 - Gemini Code Assist가 자동으로 1차 리뷰를 진행합니다
 - 추가 질문이 필요하면 `@gemini-code-assist`를 멘션하세요
 - **명령어**:
   - `/gemini review` - 새 리뷰 요청
   - `/gemini summary` - 변경사항 요약
+
