+++ .github/workflows/ci.yml
@@ -14,6 +14,12 @@ jobs:
       contents: read
       pull-requests: read
 
+    env:
+      SUPABASE_URL: ${{ secrets.SUPABASE_URL }}
+      SUPABASE_SERVICE_ROLE_KEY: ${{ secrets.SUPABASE_SERVICE_ROLE_KEY }}
+      TOUR_API_URL: ${{ secrets.TOUR_API_URL }}
+      TOUR_API_KEY: ${{ secrets.TOUR_API_KEY }}
+
     steps:
       - uses: actions/checkout@v4
         with:
@@ -28,7 +34,7 @@ jobs:
       - name: Set up Node.js
         uses: actions/setup-node@v4
         with:
-          node-version: 20
+          node-version: 22
           cache: 'pnpm'
 
       - name: Install dependencies
